Changed startPosition property of Pie diagrams to be a double (was: int) Actually make use of that property in drawing
git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/trunk/kdchart@177 c66179b2-ff0d-0410-8aef-1b1a91e65cfa

+++ examples/Pie/mainwindow.h
@@ -18,7 +18,7 @@ public:
     MainWindow( QWidget* parent = 0 );
 
 private slots:
-    void on_startPositionSB_valueChanged( int pos );
+    void on_startPositionSB_valueChanged( double pos );
 
 private:
     KDChart::Chart* m_chart;
